feat(auth): Join을 /user에서 /auth로 옮김
