+++ src/app/app.module.ts
@@ -4,9 +4,10 @@ import { AppService } from './app.service';
 import { CommonModule } from '../common/common.module';
 import { UserModule } from '../user/user.module';
 import { configModule } from './modules/config.module';
+import { AuthModule } from 'src/auth/auth.module';
 
 @Module({
-  imports: [CommonModule, UserModule, configModule],
+  imports: [CommonModule, UserModule, configModule, AuthModule],
   controllers: [AppController],
   providers: [AppService],
 })
