+++ src/auth/auth.module.ts
@@ -1,9 +1,11 @@
 import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
+import { UserService } from 'src/user/user.service';
+import { UserRepository } from 'src/user/user.repository';
 
 @Module({
+  providers: [AuthService, UserService, UserRepository],
   controllers: [AuthController],
-  providers: [AuthService]
 })
 export class AuthModule {}
